portal || add quantity logist and drivers in a company

+++ portal_service/src/main/java/ru/gnivc/portal/dto/company/CompanyDto.java
@@ -7,7 +7,7 @@ public record CompanyDto(
     String address,
     String kpp,
     String ogrn,
-    Integer quantityLogists,
-    Integer quantityDrivers
+    Long quantityLogists,
+    Long quantityDrivers
 ) {
 }
